MemSnoop

MemSnoop

Idle page tracking for memory analysis of running applications.

Security Warning

This package requires running julia as root because it needs access to several protected kernel files. To minimize your risk, this package's only dependency is the Serialization standard library. However, use this package at your own risk.

Generating a Trace

Tracking a process is easy, simply call trace

MemSnoop.trace — Function.

trace(process::AbstractProcess, measurements::NamedTuple; kw...) -> NamedTuple

Perform a measurement trace on process. The measurements to be performed are specified by the measurements argument. Return a NamedTupleTwith the same names asmeasurements` but whose values are the measurement data.

The general flow of this function is as follows:

Sleep for sampletime

Call prehook on process

Call measure on each measurement.

Call callback

Call posthook on process

Repeat for each element of iter.

Measurements

measuremts::NamedTuple : A NamedTuple where each element is some AbstractMeasurement.

Keyword Arguments

sampletime : Seconds between reading and reseting the idle page flags to determine page activity. Default: 2

iter : Iterator to control the number of samples to take. Default behavior is to keep sampling until monitored process terminates. Default: Run until program terminates.

callback : Optional callback for printing out status information (such as number of iterations).

Example

Do five measurements of idle page tracking on the julia process itself.

julia> process = MemSnoop.SnoopedProcess(getpid())
MemSnoop.SnoopedProcess{MemSnoop.Unpausable}(15703)

julia> measurements = (
    initial_timestamp = MemSnoop.Timestamp(),
    idlepages = MemSnoop.IdlePageTracker(),
    final_timestamp = MemSnoop.Timestamp(),
);

julia> data = trace(
    process,
    measurements;
    sampletime = 1,
    iter = 1:5
);

# Introspect into `data`
julia> typeof(data)
NamedTuple{(:initial_timestamp, :idlepages, :final_timestamp),Tuple{Array{Dates.DateTime,1},Array{Sample,1},Array{Dates.DateTime,1}}}

source

The trace function returns a Vector{Sample}. Each Sample contains

The VMAs assigned to the process for that sample interval.

The virtual page numbers of pages that were active during that sample interval. Internally, these are stored as a SortedRangeVector for compression.

Performing Analysis

Because of the dependency restriction, only basic functionality is provided in the MemSnoop package itself. For more detailed analyses that can be performed after a trace is generated, use the package SnoopAnalyzer. Eventually, this package will also be documented here.