Distance / WSS

Distance / WSS

To perform an analysis to estimate the Working Set Size (WSS) of an application, as well as the Memory Reuse Distance of pages, use the MemSnoop.track_distance function.

MemSnoop.track_distance — Function.

track_distance(pid; [sampletime], [iter], [filter]) -> DistanceTracker

Return a DistanceTracker with memory usage statistics for the system process with pid. Like Trace, this function will gracefully exit when pid no longer exists.

The general flow of this function is as follows:

Sleep for sampletime.

Pause pid.

Get the VMAs for pid, applying filter.

Read all of the active pages.

Mark all pages as idle.

Resume `pid.

Repeat for each element of iter.

Keyword Arguments

sampletime : Time between reading and reseting the idle page flags to determine page activity. Default: 2

iter : Iterator to control the number of samples to take. Default behavior is to keep sampling until monitored process terminates. Default: Forever()

filter : Filter to apply to process VMAs to reduce total amount of memory tracked. Default: tautology

source

MemSnoop.DistanceTracker — Type.

Struct tracking various page use metrics for an application without keeping a whole trace of page activity.

Fields

distances::Dict{Int,Int} - Reuse distance count. Keys to the dictionary are (an upper bound on) the number of unique pages accessed between subsequent accesses to an individual page. Values of the dictionary are the count of that distance.

The distances (keys) are an upper bound because of the discrete sampletime.

resident_pages::Vector{Int} - The number of resident pages. Accessing resident_pages[i] gives the number of resident pages for sample window i.

active_pages::Vector{Int} - The number of active pages.

vma_size::Vector{Int} - Collective size of all the monitored VMA regions.

max_depth::Vector{Int} - The maximum depth seen for a sample.

source

BucketStack

A custom MemSnoop.BucketStack type is used to perform the stack analysis to estimate page Reuse Distance.

MemSnoop.BucketStack — Type.

BucketStack{T}

Like a stack, but entries in the stack are buckets of elements rather than single elements. This models the fact that we sample active pages over an interval and don't have an exact trace.

Fields

buckets::Vector{Set{T}}

Constructor

BucketStack{T}()

Construct an empty BucketStack with element types T.

source

MemSnoop.upstack! — Function.

upstack!(stack::BucketStack, item) -> Int

Add item to the first bucket of stack. Delete item from lower buckets in the stack and return the depth of the item. If item was not previously found in stack, return -1.

source