Disk IO

Disk IO

SystemSnoop.DiskIO — Type.

Read from /proc/diskstats for a set of devices. Return a Dict{Symbol,DiskStats} where the keys are the devices being measured and the values are DiskStats for each device.

Fields

devices::Vector{String} - Names of devices for which to take measurements.

Constructor

SystemSnoop.DiskIO(devices) -> DiskIO

Documentation on /proc/diskstats

The /proc/diskstats file displays the I/O statistics of block devices. Each line contains the following 14 fields:

 1 - major number
 2 - minor mumber
 3 - device name
 4 - reads completed successfully
 5 - reads merged
 6 - sectors read
 7 - time spent reading (ms)
 8 - writes completed
 9 - writes merged
10 - sectors written
11 - time spent writing (ms)
12 - I/Os currently in progress
13 - time spent doing I/Os (ms)
14 - weighted time spent doing I/Os (ms)

Kernel 4.18+ appends four more fields for discard tracking putting the total at 18:

15 - discards completed successfully
16 - discards merged
17 - sectors discarded
18 - time spent discarding

source

SystemSnoop.DiskStats — Type.

Storate for DiskIO.

Fields

reads_completed - Reads successfully completed

reads_merged

sectors_read

time_reading (units: ms). Note that this field is for all pending operations, and will add time for multiple read requests.

writes_completed

writes_merged

sectors_written

time_writing (units: ms). Note that this field is for all pending operations, and will add time for multiple read requests.

time_io (units: ms). Note, this is wall for the total time this disk was busy.

source