Idle Page Tracking

Idle Page Tracking

MemSnoop.IdlePageTracker — Type.

Measurement type for performing Idle Page Tracking on a process. To filter process VMAs, construct as

IdlePageTracker([filter])

where filter is a VMA filter function.

Implementation Details

filter - The VMA filter to apply. Defaults to all VMAs.

vmas::Vector{VMA} - Buffer for storing VMAs.

buffer::Vector{UInt64} - Buffer to store the idle page bitmap.

source

MemSnoop.initbuffer! — Method.

initbuffer!(I::IdlePageTracker)

Read once from page_idle/bitmap to get the size of the bitmap. Set the bitmap in I to this size to avoid reallocation every time the bitmap is read.

source

MemSnoop.markidle — Method.

markidle(pid, vmas)

Mark all of the memory pages in the list of vmas for process with pid as idle.

source

MemSnoop.readidle — Method.

readidle(pid, vmas, buffer) -> SortedRangeVector{UInt}

Return the active pages within vmas of process with pid. Use buffer as storage for the idle bitmap to avoid allocations

source

MemSnoop.walkpagemap — Method.

walkpagemap(f::Function, pid, vmas; [buffer::Vector{UInt64}])

For each VMA in iterator vmas, store the contents of /proc/pid/pagemap into buffer for this VMA and call f(buffer).

Note that it is possible for buffer to be empty.

source