Sample

Sample

SystemSnoop.Sample — Type.

Simple container containing the list of VMAs analyzed for a sample as well as the individual pages accessed.

Fields

vmas :: Vector{VMA} - The VMAs analyzed during this sample.

pages :: SortedRangeVector{UInt64} - The pages that were active during this sample. Pages are encoded by virtual page number. To get an address, multiply the page number by the pagesize (generally 4096).

Methods

vmas - VMAs of Sample.

pages - Active pages from Sample or Vector{Sample}.

wss - Working set size of Sample.

union - Merge two Samples together.

isactive - Check if a page was active in Sample.

bitmap - Construct a bitmap of active pages for a Vector{Sample}.

source

SystemSnoop.bitmap — Method.

bitmap(trace::Vector{Sample}, vma::VMA) -> Array{Bool, 2}

Return a bitmap B of active pages in trace with virtual addresses from vma. B[i,j] == true if the ith address in vma in trace[j] is active.

source

SystemSnoop.isactive — Method.

isactive(sample::Sample, page) -> Bool

Return true if page was active in sample.

source

SystemSnoop.pages — Method.

pages(trace::Vector{Sample}) -> Vector{UInt64}

Return a sorted vector of all pages in trace that were marked as "active" at least once. Pages are encoded by virtual page number.

source

SystemSnoop.pages — Method.

pages(sample::Sample) -> Set{UInt64}

Return a set of all active pages in sample.

source

SystemSnoop.vmas — Method.

vmas(trace::Vector{Sample}) -> Vector{VMA}

Return the largest sorted collection $V$ of VMAs with the property that for any sample $S \in trace$ and for any VMA $s \in S$, $s \subset v$ for some $v \in V$ and $s \cap u = \emptyset$ for all $u \in V \setminus v$.o

source

SystemSnoop.wss — Method.

wss(S::Sample) -> Int

Return the number of active pages for S.

source