Utility Functions

Utility Functions

MemSnoop.inmemory — Method.

inmemory(x::UInt) -> Bool

Return true if x (interpreted as an entry in Linux /pagemap) if located in memory.

source

MemSnoop.isactive — Method.

isactive(x::Integer, buffer::Vector{UInt64}) -> Bool

Return true if bit x of buffer is set, intrerpreting buffer as a contiguous chunk of memory.

source

MemSnoop.isbitset — Method.

isbitset(x::Integer, b::Integer) -> Bool

Return true if bit b of x is 1.

source

MemSnoop.pfnmask — Method.

pfnmask(x::UInt) -> UInt

Return the lower 55 bits of x. When applied to a /proc/pid/pagemap entry, returns the physical page number (pfn) of that entry.

source