VMAs

VMAs

SystemSnoop.VMA — Type.

Translated Virtual Memory Area (VMA) for a process.

Fields

start::UInt64 - The starting virtual page number for the VMA.

stop::UInt64 - The last valid virtual page number for the VMA.

remainder::String - The remainder of the entry in /proc/pid/maps.

Methods

length, startaddress, stopaddress

Filter Functions

There are a handful of builtin filter functions to help get rid of unwanted VMAs.

heap

readable

writable

executable

flagset

longerthan

source

Base.compact — Method.

compact(vmas::Vector{VMA}) -> Vector{VMA}

Given an unsorted collection vmas, return the smallest collection $V$ such that

For any $u \in vmas$, $u \subset v$ for some $v \in V$.

All elements of $V$ are disjoint.

$V$ is sorted by starting address.

source

Base.issubset — Method.

issubset(a::VMA, b::VMA) -> Bool

Return true if VMA region a is a subset of b.

source

Base.length — Method.

length(vma::VMA) -> Int

Return the size of vma in number of pages.

source

Base.union — Method.

union(a::VMA, b::VMA) -> VMA

Return a VMA that is the union of the regions covered by a and b. Assumes that a and b are overlapping.

source

SystemSnoop.executable — Method.

Return true if vma is executable.

source

SystemSnoop.flagset — Method.

Return true if vma is either readable, writeable, or executable

source

SystemSnoop.getvmas! — Function.

getvmas!(buffer::Vector{VMA}, pid, [filter])

Fill buffer with the Virtual Memory Areas associated with the process with pid. Can optinally supply a filter. VMAs in buffer will be sorted by virtual address.

Filter

The filter must be of the form

f(vma::VMA) -> Bool

where vma is the parsed VMA region from a line of the process's maps file.

For example, if an entry in the maps file is

0088f000-010fe000 rw-p 00000000 00:00 0

then vma = VMA(0x0088f000,0x010fe000, rw-p 00000000 00:00 0)

source

SystemSnoop.heap — Method.

Return true if vma is for the heap.

source

SystemSnoop.longerthan — Method.

longerthan(x, n) -> Bool

Return true if length(x) > n

longerthan(n) -> Function

Return a function x -> longerthan(x, n)

source

SystemSnoop.overlapping — Method.

overlapping(a::VMA, b::VMA) -> Bool

Return true if VMA regions a and b overlap.

source

SystemSnoop.readable — Method.

Return true if vma is readable.

source

SystemSnoop.startaddress — Method.

startaddress(vma::VMA) -> UInt

Return the first virtual addresses assigned to vma.

source

SystemSnoop.stopaddress — Method.

stopaddres(vma::VMA) -> UInt

Return the last virtual addresses assigned to vma.

source

SystemSnoop.writable — Method.

Return true if vma is writable.

source