Process IO

Process IO

SystemSnoop.BytesIO — Type.

Struct representing measurements for ProcessIO

Fields

rchar - Characters Read. The number of bytes which this task has caused to be read from storage. This is simply the sum of bytes which this process passed to read(2) and similar system calls. It includes things such as terminal I/O and is unaffected by whether or not actual physical disk I/O was required (the read might have been satisfied from pagecache).

wchar - Characters Written. The number of bytes which this task has caused, or shall cause to be written to disk. Similar caveats apply here as with rchar.

read_bytes - Attempt to count the number of bytes which this process really did cause to be fetched from the storage layer. This is accurate for block-backed filesystems.

write_bytes - Attempt to count the number of bytes which this process caused to be sent to the storage layer.

source

SystemSnoop.ProcessIO — Type.

Record the rchar, wchar, read_bytes, and write_bytes fields of /proc/pid/io.

Each measurement returns a BytesIO object.

Example read from /proc/pid/io:

rchar: 3089822
wchar: 139463
syscr: 159
syscw: 178
read_bytes: 0
write_bytes: 4096
cancelled_write_bytes: 0

becomes

BytesIO(3089822, 139463, 0, 4096)

source