Process

Process

MemSnoop.SnoopedProcess — Type.

Struct container a pid as well as auxiliary data structure to make the snooping process more efficient.

Fields

pid::Int64 - The pid of the process.

Constructor

SnoopedProcess(pid) -> SnoopedProcess

Construct a Process with the given pid.

Methods

getpid - Get the PID of this process.

prehook - Method to call before measurements.

posthook - Method to call after measurements.

source

MemSnoop.posthook — Method.

posthook(P::AbstractProcess)

If P is a pausable process, unpause P.

source

MemSnoop.prehook — Method.

prehook(P::AbstractProcess)

If P is a pausable process, pause P.

source