Sorted Range Vectors

Sorted Range Vectors

MemSnoop.SortedRangeVector — Type.

Compact representation of data of type T that is both sorted and usually occurs in contiguous ranges. For example, since groups of virtual memory pages are usually accessed together, a SortedRangeVector can encode those more compactly than a normal vector.

Fields

ranges :: Vector{UnitRange{T} - The elements of the SortedRangeVector, compacted into contiguous ranges.

Constructor

SortedRangeVector{T}() -> SortedRangeVector{T}

Construct a empty SortedRangeVector with element type T.

Methods

sumall

lastelement

push!

in

union

source

MemSnoop.lastelement — Method.

lastelement(V::SortedRangeVector{T}) -> T

Return the last element of the last range of V.

source

MemSnoop.sumall — Method.

sumall(V::SortedRangeVector)

Return the sum of lengths of each element of V.

source

Base.in — Method.

in(x, V::SortedRangeVector) -> Bool

Perform an efficient search in V for x.

source

Base.push! — Method.

push!(V::SortedRangeVector{T}, x::T) where {T}

Add x to the end of V, merging x into the final range if appropriate.

push!(V::SortedRangeVector{T}, x::UnitRange{T}) where {T}

Merge x with the final range in V if they overlap. Otherwise, append x to the end of V.

NOTE: Assumes that first(x) >= first(last(V))

source

Base.union — Method.

union(A::SortedRangeVector, B::SortedRangeVector)

Efficiently union A and B together.

source