MemSnoop

MemSnoop

Idle page tracking for memory analysis of running applications.

Security Warning

This package requires running julia as root because it needs access to several protected kernel files. To minimize your risk, this package's only dependency is the Serialization standard library. However, use this package at your own risk.

Generating a Trace

Tracking a process is easy, simply call trace

MemSnoop.trace — Function.

trace(pid; [sampletime], [iter], [filter], [callback]) -> Vector{Sample}

Record the full trace of pages accessed by an application with pid. Function will gracefully exit and return Vector{Sample} if process pid no longer exists.

The general flow of this function is as follows:

Sleep for sampletime.

Pause pid.

Get the VMAs for pid, applying filter.

Read all of the active pages.

Mark all pages as idle.

Call callback

Resume `pid.

Repeat for each element of iter.

Keyword Arguments

sampletime : Seconds between reading and reseting the idle page flags to determine page activity. Default: 2

iter : Iterator to control the number of samples to take. Default behavior is to keep sampling until monitored process terminates. Default: Run until program terminates.

filter : Filter to apply to process VMAs to reduce total amount of memory tracked.

callback : Optional callback for printing out status information (such as number of iterations).

source

The trace function returns a Vector{Sample}. Each Sample contains

The VMAs assigned to the process for that sample interval.

The virtual page numbers of pages that were active during that sample interval. Internally, these are stored as a SortedRangeVector for compression.

Performing Analysis

Because of the dependency restriction, only basic functionality is provided in the MemSnoop package itself. For more detailed analyses that can be performed after a trace is generated, use the package SnoopAnalyzer. Eventually, this package will also be documented here.