Process

Process

MemSnoop.Process — Type.

Struct container a pid as well as auxiliary data structure to make the snooping process more efficient.

Fields

pid::Int64 - The pid of the process.

vmas::Vector{VMA} - Buffer for storing the VMAs assigned to this process.

buffer::Vector{UInt64} - Auxiliary array used as a buffer for the idle page map.

Constructor

Process(pid) -> Process

Construct a Process with the given pid.

Methods

initbuffer!

getvmas!

walkpagemap

markidle

readidle

source

MemSnoop.initbuffer! — Method.

initbuffer!(p::AbstractProcess)

Read once from page_idle/bitmap to get the size of the bitmap. Set the bitmap in p to this size to avoid reallocation every time the bitmap is read.

source

MemSnoop.markidle — Method.

markidle(process::AbstractProcess)

Mark all of the memory pages assigned to process as idle.

source

MemSnoop.readidle — Method.

readidle(process::AbstractProcess) -> SortedRangeVector{UInt}

Return the active pages of process.

source

MemSnoop.walkpagemap — Method.

walkpagemap(f::Function, pid, vmas; [buffer::Vector{UInt64}])

For each VMA in iterator vmas, store the contents of /proc/pid/pagemap into buffer for this VMA and call f(buffer).

Note that it is possible for buffer to be empty.

source