Process

Process

SystemSnoop.SnoopedProcess — Type.

Struct container a pid as well as auxiliary data structure to make the snooping process more efficient. SnoopedProcesses come in two variants, Pausable and Unpausable.

Pausable processes will be paused before a set of measurements are taken by calling kill -STOP and resumed after afterwards by calling kill -CONT. Unpausable processes will not be touched.

To construct a Pausable process with pid, call

ps = SnoopedProcess{Pausable}(pid)

To construct an Unpausable process, call

ps = SnoopedProcess{Unpausable}(pid)

Fields

pid::Int64 - The pid of the process.

Methods

getpid - Get the PID of this process.

isrunning - Return true if process is running.

prehook - Method to call before measurements.

posthook - Method to call after measurements.

source

SystemSnoop.posthook — Method.

posthook(P::AbstractProcess)

If P is a pausable process, unpause P.

source

SystemSnoop.prehook — Method.

prehook(P::AbstractProcess)

If P is a pausable process, pause P.

source