Generating Traces

Generating Traces

MemSnoop has the ability to record the full trace of pages accessed by an application. This is performed using trace

MemSnoop.trace — Function.

trace(pid; [sampletime], [iter], [filter]) -> Vector{Sample}

Record the full trace of pages accessed by an application with pid. Function will gracefully exit and return Vector{Sample} if process pid no longer exists.

The general flow of this function is as follows:

Sleep for sampletime.

Pause pid.

Get the VMAs for pid, applying filter.

Read all of the active pages.

Mark all pages as idle.

Resume `pid.

Repeat for each element of iter.

Keyword Arguments

sampletime : Seconds between reading and reseting the idle page flags to determine page activity. Default: 2

iter : Iterator to control the number of samples to take. Default behavior is to keep sampling until monitored process terminates. Default: Run until program terminates.

filter : Filter to apply to process VMAs to reduce total amount of memory tracked.

source

MemSnoop.pages — Method.

pages(trace::Vector{Sample}) -> Vector{UInt64}

Return a sorted vector of all pages in trace that were marked as "active" at least once. Pages are encoded by virtual page number.

source

MemSnoop.vmas — Method.

vmas(trace::Vector{Sample}) -> Vector{VMA}

Return the largest sorted collection $V$ of VMAs with the property that for any sample $S \in trace$ and for any VMA $s \in S$, $s \subset v$ for some $v \in V$ and $s \cap u = \emptyset$ for all $u \in V \setminus v$.o

source

Implementation Details - Sample

MemSnoop.Sample — Type.

Simple container containing the list of VMAs analyzed for a sample as well as the individual pages accessed.

Fields

vmas :: Vector{VMA} - The VMAs analyzed during this sample.

pages :: SortedRangeVector{UInt64} - The pages that were active during this sample. Pages are encoded by virtual page number. To get an address, multiply the page number by the pagesize (generally 4096).

source

MemSnoop.isactive — Method.

isactive(sample::Sample, page) -> Bool

Return true if page was active in sample.

source

MemSnoop.pages — Method.

pages(sample::Sample) -> Set{UInt64}

Return a set of all active pages in sample.

source

MemSnoop.bitmap — Method.

bitmap(trace::Vector{Sample}, vma::VMA) -> Array{Bool, 2}

Return a bitmap B of active pages in trace with virtual addresses from vma. B[i,j] == true if the ith address in vma in trace[j] is active.

source

Implementation Details - SortedRangeVector

Since pages are generally accessed sequentially, the record of active pages is encoded as a MemSnoop.SortedRangeVector that compresses contiguous runs of accesses. Note that there is an implicit assumption that the VMAs are ordered, which should be the case since /prod/pid/maps orderes VMAs.

MemSnoop.SortedRangeVector
MemSnoop.lastelement
push!(::MemSnoop.SortedRangeVector{T}, x::T) where T
MemSnoop.insorted