Traces

Traces

MemSnoop.trace — Function.

trace(process::AbstractProcess, measurements::Tuple; kw...) -> Tuple

Perform a measurement trace on process. The measurements to be performed are specified by the measurements argument. Return a tuple T where T[i] is the data for measurements[i].

The general flow of this function is as follows:

Sleep for sampletime

Call prehook on process

Perform measurements

Call callback

Call posthook on process

Repeat for each element of iter.

Measurements

measuremts::Tuple : A tuple where each element is some AbstractMeasurement. By default, the only measurement performed is IdlePageTracker.

Keyword Arguments

sampletime : Seconds between reading and reseting the idle page flags to determine page activity. Default: 2

iter : Iterator to control the number of samples to take. Default behavior is to keep sampling until monitored process terminates. Default: Run until program terminates.

callback : Optional callback for printing out status information (such as number of iterations).

source

MemSnoop.AbstractMeasurement — Type.

Abstract supertype for process measurements.

Required API

prepare

measure

Optional API

initialize!

Concrete Implementations

Timestamp

IdlePageTracker

source

MemSnoop.Timestamp — Type.

Collect timestamps.

source

MemSnoop.initialize! — Method.

initialize!(M::AbstractMeasurement, process::AbstractProcess)

Perform any initialization required for measurement M. This method is optional and defaults to a no-op.

source

MemSnoop.measure — Method.

measure(M::AbstractMeasurement, process::AbstractProcess) -> T

Return data of type T.

source

MemSnoop.prepare — Method.

prepare(M::AbstractMeasurement) -> Vector{T}

Return an empty vector to hold measurement data of type T for measurement M.

source