Traces

Traces

MemSnoop.trace — Method.

trace(process::AbstractProcess, measurements::NamedTuple; kw...) -> NamedTuple

Perform a measurement trace on process. The measurements to be performed are specified by the measurements argument. The values of this tuple are AbstractMeasurements.

Return a NamedTuple T with the same names as measurements but whose values are the measurement data.

The general flow of this function is as follows:

Sleep for sampletime

Call prehook on process

Call measure on each measurement.

Call callback

Call posthook on process

Repeat for each element of iter.

Measurements

measurements::NamedTuple : A NamedTuple where each element is some AbstractMeasurement.

Keyword Arguments

sampletime : Seconds between reading and reseting the idle page flags to determine page activity. Default: 2

iter : Iterator to control the number of samples to take. Default behavior is to keep sampling until monitored process terminates. Default: Run until program terminates.

callback : Optional callback for printing out status information (such as number of iterations).

Example

Do five measurements of idle page tracking on the julia process itself.

julia> process = MemSnoop.SnoopedProcess(getpid())
MemSnoop.SnoopedProcess{MemSnoop.Unpausable}(15703)

julia> measurements = (
    initial_timestamp = MemSnoop.Timestamp(),
    idlepages = MemSnoop.IdlePageTracker(),
    final_timestamp = MemSnoop.Timestamp(),
);

julia> data = trace(
    process,
    measurements;
    sampletime = 1,
    iter = 1:5
);

# Introspect into `data`
julia> typeof(data)
NamedTuple{(:initial_timestamp, :idlepages, :final_timestamp),Tuple{Array{Dates.DateTime,1},Array{Sample,1},Array{Dates.DateTime,1}}}

See also: AbstractMeasurement, SnoopedProcess

source

MemSnoop.AbstractMeasurement — Type.

Abstract supertype for process measurements.

Required API

prepare

measure

Concrete Implementations

Timestamp

IdlePageTracker

DiskIO

Statm

source

MemSnoop.Timestamp — Type.

Collect timestamps.

source

MemSnoop.measure — Method.

measure(M::AbstractMeasurement, P::AbstractProcess) -> T

Return data of type T.

source

MemSnoop.prepare — Method.

prepare(M::AbstractMeasurement, P::AbstractProcess) -> Vector{T}

Return an empty vector to hold measurement data of type T for measurement M. Any initialization required M should happen here.

source