Traces

Traces

SystemSnoop.trace — Method.

trace(process, measurements::NamedTuple; kw...) -> NamedTuple

Perform a measurement trace on process. The measurements to be performed are specified by the measurements argument. The values of this tuple are types that implement the prepare and measure interface.

Return a NamedTuple T with the same names as measurements but whose values are the measurement data.

Argument process can be:

A SnoopedProcess

An integer representing a process PID

A Base.Process spawned by Julia

A Cmd that will launch a process

The general flow of this function is as follows:

Sleep for sampletime

Call prehook on process

Call measure on each measurement.

Call callback

Call posthook on process

Repeat for each element of iter.

Measurements

measurements::NamedTuple : A NamedTuple where each element implements prepare and measure.

Keyword Arguments

sampletime : Seconds between reading and reseting the idle page flags to determine page activity. Can also pass a SmartSample for better control of sample times. Default: 2

iter : Iterator to control the number of samples to take. Default behavior is to keep sampling until monitored process terminates. Default: Run until program terminates.

callback : Optional callback for printing out status information (such as number of iterations).

Example

Do five measurements of idle page tracking on the top command.

julia> measurements = (
    initial_timestamp = SystemSnoop.Timestamp(),
    idlepages = SystemSnoop.IdlePageTracker(),
    final_timestamp = SystemSnoop.Timestamp(),
);

julia> data = trace(
    `top`,
    measurements;
    sampletime = 1,
    iter = 1:5
);

# Introspect into `data`
julia> typeof(data)
NamedTuple{(:initial_timestamp, :idlepages, :final_timestamp),Tuple{Array{Dates.DateTime,1},Array{Sample,1},Array{Dates.DateTime,1}}}

See also: SnoopedProcess, SmartSample

source

SystemSnoop.SmartSample — Type.

SystemSnoop.SmartSample(t::TimePeriod) -> SmartSample

Smart Sampler to ensure measurements happen every t time units. Samples will happen at multiples of t from the first measurement. If a sample period is missed, the sampler will wait until the next appropriate multiple of t.

source

SystemSnoop.Timestamp — Type.

Collect timestamps.

source