Utilities

Utilities

MemSnoop.Forever — Type.

In iterator that returns an infinite amount of nothing.

source

MemSnoop.PIDException — Type.

Exception indicating that process with pid no longer exists.

source

MemSnoop.Timeout — Type.

An iterator that stops iterating at a certain time.

Example

julia> using Dates

julia> y = MemSnoop.Timeout(Second(10))
MemSnoop.Timeout(2019-01-04T11:51:09.262)

julia> now()
2019-01-04T11:50:59.274

julia> for i in y; end

julia> now()
2019-01-04T11:51:09.275

source

MemSnoop.increment! — Method.

increment!(d::AbstractDict, k, v)

Increment d[k] by v. If d[k] does not exist, initialize it to v.

source

MemSnoop.isrunning — Method.

isrunning(pid) -> Bool

Return true is a process with pid is running.

source

MemSnoop.pause — Method.

pause(pid)

Pause process with pid. If process does not exist, throw a PIDException.

source

MemSnoop.pidsafeopen — Method.

pidfraceopen(f::Function, file::String, pid, args...; kw...)

Open system pseudo file file for process with pid and pass the handle to f. If a File does not exist error is thown, throws a PIDException instead.

Arguments args and kw are forwarded to the call to open.

source

MemSnoop.resume — Method.

resume(pid)

Resume process with pid. If process does not exist, throw a PIDException

source

MemSnoop.safeparse — Method.

safeparse(::Type{T}, str; base = 10) -> T

Try to parse str to type T. If that fails, return zero(T).

source