Utilities

Utilities

MemSnoop.Forever — Type.

In iterator that returns an infinite amount of nothing.

source

MemSnoop.PIDException — Type.

Exception indicating that process with pid no longer exists.

source

MemSnoop.inmemory — Method.

inmemory(x::UInt) -> Bool

Return true if x (interpreted as an entry in Linux /pagemap) if located in memory.

source

MemSnoop.isactive — Method.

isactive(x::Integer, buffer::Vector{UInt64}) -> Bool

Return true if bit x of buffer is set, intrerpreting buffer as a contiguous chunk of memory.

source

MemSnoop.isbitset — Method.

isbitset(x::Integer, b::Integer) -> Bool

Return true if bit b of x is 1.

source

MemSnoop.pause — Method.

pause(pid)

Pause process with pid. If process does not exist, throw a PIDException.

source

MemSnoop.pfnmask — Method.

pfnmask(x::UInt) -> UInt

Return the lower 55 bits of x. When applied to a /proc/pid/pagemap entry, returns the physical page number (pfn) of that entry.

source

MemSnoop.resume — Method.

resume(pid)

Resume process with pid. If process does not exist, throw a PIDException

source