VMAs

VMAs

Internally, MemSnoop works on VMA boundaries, allowing for filtering on VMAs, and for merging overlapping VMAs together. The latter is important when aggregating data across multiple samples, as various VMAs may change in size from sample to sample.

Type

MemSnoop.VMA — Type.

Translated Virtual Memory Area (VMA) for a process.

Fields

start::UInt64 - The starting virtual page number for the VMA.

stop::UInt64 - The last valid virtual page number for the VMA.

remainder::String - The remainder of the entry in /proc/pid/maps.

Methods

length, translate, startaddress, stopaddress

source

Methods

Basic Functions

MemSnoop.startaddress — Method.

startaddress(vma::VMA) -> UInt

Return the first virtual addresses assigned to vma.

source

MemSnoop.stopaddress — Method.

stopaddres(vma::VMA) -> UInt

Return the last virtual addresses assigned to vma.

source

Base.length — Method.

length(vma::VMA) -> Int

Return the size of vma in number of pages.

source

Set-like Functions

MemSnoop.overlapping — Function.

overlapping(a::VMA, b::VMA) -> Bool

Return true if VMA regions a and b overlap.

source

Base.issubset — Method.

issubset(a::VMA, b::VMA) -> Bool

Return true if VMA region a is a subset of b.

source

Base.union — Method.

union(a::VMA, b::VMA) -> VMA

Return a VMA that is the union of the regions covered by a and b. Assumes that a and b are overlapping.

source

Base.compact — Function.

compact(vmas::Vector{VMA}) -> Vector{VMA}

Given an unsorted collection vmas, return the smallest collection $V$ such that

For any $u \in vmas$, $u \subset v$ for some $v \in V$.

All elements of $V$ are disjoint.

$V$ is sorted by starting address.

source

Filters

Filters can be applied to VMAs to prune cut down on the amount of data collected. Built-in filters are described below.

MemSnoop.heap — Function.

Return true if vma is for the heap.

source

MemSnoop.readable — Function.

Return true if vma is readable.

source

MemSnoop.writable — Function.

Return true if vma is writable.

source

MemSnoop.executable — Function.

Return true if vma is executable.

source

MemSnoop.flagset — Function.

Return true if vma is either readable, writeable, or executable

source

MemSnoop.longerthan — Function.

longerthan(x, n) -> Bool

Return true if length(x) > n

longerthan(n) -> Function

Return a function x -> longerthan(x, n)

source